Introducing VfmdLineArray and VfmdLineArrayIterator
VfmdLineArray is just an array of VfmdLine objects.
VfmdLineArrayIterator lets you iterate over the VfmdLineArray
in chunks of lines or bytes or patterns or however. This helps
us process the VfmdLineArray as a bytestream, which is needed
to identify the span tags in the bytestream.

So, effectively, we want to be able to identify the span tags
in an array of lines without having to merge the lines into one
large continuous bytearray (so we can save on copying time).