VfmdPreprocessor: Simplify the preprocessor
The previous UTF-8 preprocessing code was a heavily modified version
of pcre_valid_utf8.c because it had to buffer the output
(when the data provided was not enough to complete a line's
worth of output).

This commit makes the preprocessor buffer in the input side.
So, the code is only slightly modified from pcre_valid_utf8.c,
and is therefore more reliable.